added dodge while class is archer

+++ RPGTextGame/Characters.cs
@@ -6,7 +6,7 @@ public class Characters
     public virtual int Attack {get; set;}
     public virtual int Defence {get; set;}
     public virtual int Intelligence {get; set;}
-    public virtual double DodgeChance {get; set;}
+    public virtual int DodgeChance {get; set;}
     
     public int GetDamage(Characters enemy, bool isDefensiveStance)
     {
